Add dmtc0 and context write

+++ src/n64/headers/n64/cpu.h
@@ -39,7 +39,7 @@ void write_call(N64& n64, u64 pc);
 
 void cycle_tick(N64 &n64, u32 cycles);
 
-void write_cop0(N64 &n64, u64 v, u32 reg);
+void write_cop0(N64 &n64, u64 v, u64 reg);
 u64 read_cop0(N64& n64, u32 reg);
